+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - #51: Improve Development Environment and upgrade Nautobot version to 1.1.0
 
+### Fixed
+
+- #53: The **SINCE** filter to receive email notifications is extended on day in the past in order to get notifications from the same day as the last notifications stored.
+
 ## v0.1.7 - 2021-07-27
 
 ### Added